Move post-detail-data controller back into directive
